+++ src/test/java/com/lazerycode/jmeter/configuration/RemoteArgumentsArrayBuilderTest.java
@@ -1,63 +1,65 @@
 package com.lazerycode.jmeter.configuration;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import com.lazerycode.jmeter.properties.ConfigurationFiles;
+import com.lazerycode.jmeter.properties.PropertiesFiles;
+import com.lazerycode.jmeter.properties.PropertiesMapping;
+import org.junit.Before;
+import org.junit.Test;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 
-import org.junit.Before;
-import org.junit.Test;
-
-import com.lazerycode.jmeter.properties.JMeterPropertiesFiles;
-import com.lazerycode.jmeter.properties.PropertyContainer;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 
 public class RemoteArgumentsArrayBuilderTest {
 
-	private Map<JMeterPropertiesFiles, PropertyContainer> inputMap = new HashMap<JMeterPropertiesFiles, PropertyContainer>(); 
-	private RemoteArgumentsArrayBuilder arrayBuilder = new RemoteArgumentsArrayBuilder() ; 
-	private PropertyContainer container = new PropertyContainer(); 
-	
+	private Map<ConfigurationFiles, PropertiesMapping> inputMap = new HashMap<>();
+	private RemoteArgumentsArrayBuilder arrayBuilder = new RemoteArgumentsArrayBuilder();
+	private PropertiesMapping propertiesMapping;
+
 	@Before
-	public void setup() { 
-		Properties props = new Properties(); 
-		props.setProperty("hello","world"); 
-		container.setFinalPropertyObject(props); 
+	public void setup() {
+		Map<String, String> additionalProperties = new HashMap<>();
+		additionalProperties.put("hello", "world");
+		propertiesMapping = new PropertiesMapping(additionalProperties);
+		PropertiesFiles propertiesFile = new PropertiesFiles();
+		propertiesFile.addAndOverwriteProperties(additionalProperties);
+		propertiesMapping.setPropertiesFile(propertiesFile);
 	}
-	
+
 	@Test
 	public void shouldReturnEmptyListWhenNoPropsPassed() {
-		List<String> result = arrayBuilder.buildRemoteArgumentsArray(null); 
-		assertTrue(result.isEmpty()); 
+		List<String> result = arrayBuilder.buildRemoteArgumentsArray(null);
+		assertTrue(result.isEmpty());
 	}
 
 	@Test
-	public void shouldBuildCommandLineArgumentsForSystemProperties() {	
+	public void shouldBuildCommandLineArgumentsForSystemProperties() {
 
-		inputMap.put(JMeterPropertiesFiles.SYSTEM_PROPERTIES, container);	
-		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap); 
-		assertEquals(2,result.size());
-		assertEquals("-Dhello",result.get(0)); 
-		assertEquals("world",result.get(1));
+		inputMap.put(ConfigurationFiles.SYSTEM_PROPERTIES, propertiesMapping);
+		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap);
+		assertEquals(2, result.size());
+		assertEquals("-Dhello", result.get(0));
+		assertEquals("world", result.get(1));
 	}
-	
+
 	@Test
 	public void shouldBuildCommandLineArgumentsGlobalProperties() {
 
-		inputMap.put(JMeterPropertiesFiles.GLOBAL_PROPERTIES, container);	
-		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap); 
-		assertEquals(1,result.size());
-		assertEquals("-Ghello=world",result.get(0)); 
+		inputMap.put(ConfigurationFiles.GLOBAL_PROPERTIES, propertiesMapping);
+		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap);
+		assertEquals(1, result.size());
+		assertEquals("-Ghello=world", result.get(0));
 	}
-	
+
 	@Test
 	public void shoulIgnoreOtherTypesOfProperties() {
 
-		inputMap.put(JMeterPropertiesFiles.JMETER_PROPERTIES, container);	
-		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap); 
-		assertTrue(result.isEmpty()); 
+		inputMap.put(ConfigurationFiles.JMETER_PROPERTIES, propertiesMapping);
+		List<String> result = arrayBuilder.buildRemoteArgumentsArray(inputMap);
+		assertTrue(result.isEmpty());
 	}
 }