+++ src/test/java/com/lazerycode/jmeter/properties/ConfigurationFilesTest.java
@@ -6,23 +6,23 @@ import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.junit.Assert.assertThat;
 
-public class JMeterPropertiesFilesTest {
+public class ConfigurationFilesTest {
 
 	@Test
 	public void getPropertiesFileNameReturnsExpectedFilename() {
-		assertThat(JMeterPropertiesFiles.JMETER_PROPERTIES.getFilename(),
+		assertThat(ConfigurationFiles.JMETER_PROPERTIES.getFilename(),
 				is(equalTo("jmeter.properties")));
 	}
 
 	@Test
 	public void createFileIfItDoesNotExistReturnsTrue() {
-		assertThat(JMeterPropertiesFiles.JMETER_PROPERTIES.createFileIfItDoesNotExist(),
+		assertThat(ConfigurationFiles.JMETER_PROPERTIES.createFileIfItDoesNotExist(),
 				is(equalTo(true)));
 	}
 
 	@Test
 	public void createFileIfItDoesNotExistReturnsFalse() {
-		assertThat(JMeterPropertiesFiles.GLOBAL_PROPERTIES.createFileIfItDoesNotExist(),
+		assertThat(ConfigurationFiles.GLOBAL_PROPERTIES.createFileIfItDoesNotExist(),
 				is(equalTo(false)));
 	}
 }